+++ frontend/src/__test__/mocks/VotingStatisticListMock.ts
@@ -1,0 +1,23 @@
+export interface IData {
+	confirmed: {
+		total: 124252799;
+		totalPercentage: 99.99;
+	};
+	abstention: {
+		total: 124252799;
+		totalPercentage: 89.85;
+	};
+}
+
+const MOCKED_DATA: IData = {
+	confirmed: {
+		total: 124252799,
+		totalPercentage: 99.99,
+	},
+	abstention: {
+		total: 124252799,
+		totalPercentage: 89.85,
+	},
+};
+
+export default MOCKED_DATA;